requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.9.2

# PDF parsing
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
tabula-py==2.10.0

# OCR / image processing
pytesseract==0.3.13
Pillow==10.4.0
opencv-python==4.10.0.84

# Embeddings & retrieval
sentence-transformers==3.0.1
scikit-learn==1.5.1
faiss-cpu==1.8.0.post1

# Utilities
matplotlib==3.9.0
jupyter==1.0.0
rapidfuzz==3.9.6
